requirements.txt
mlx>=0.0.4
mlx-lm>=0.1.1
torch>=2.1.0
transformers>=4.39.0
huggingface_hub>=0.22.0
datasets>=2.18.0
scikit-learn>=1.3.0
numpy>=1.26.0
pyyaml>=6.0.1
tqdm>=4.66.0
sentencepiece>=0.1.99
safetensors>=0.4.2
psutil>=5.9.0
